requirements.txt
Django==1.9
git+https://github.com/django-ckeditor/django-ckeditor.git@master
git+https://github.com/zefj/django-simple-search2.git@master
pyenchant==1.6.6
requests==2.9.1
wheel==0.24.0